requirements.txt
fastapi
uvicorn
sqlalchemy
mysql-connector-python
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai
jinja2
python-dotenv
requests
Pillow
